954: Renamed Overlay to ModalOverlay
