+++ src/components/@widgets/MakeWidget/MakeWidget.tsx
@@ -59,7 +59,7 @@ import MakeOrderReview from "../../@reviewScreens/MakeOrderReview/MakeOrderRevie
 import WrapReview from "../../@reviewScreens/WrapReview/WrapReview";
 import { SelectOption } from "../../Dropdown/Dropdown";
 import OrderTypesModal from "../../InformationModals/subcomponents/OrderTypesModal/OrderTypesModal";
-import Overlay from "../../Overlay/Overlay";
+import ModalOverlay from "../../ModalOverlay/ModalOverlay";
 import ProtocolFeeOverlay from "../../ProtocolFeeOverlay/ProtocolFeeOverlay";
 import { notifyOrderCreated } from "../../Toasts/ToastController";
 import TokenList from "../../TokenList/TokenList";
@@ -534,7 +534,7 @@ const MakeWidget: FC = () => {
           />
         )}
 
-      <Overlay
+      <ModalOverlay
         hasDynamicHeight
         onClose={() => setShowTokenSelectModal(null)}
         title={t("common.selectToken")}
@@ -550,14 +550,14 @@ const MakeWidget: FC = () => {
             setShowTokenSelectModal(null);
           }}
         />
-      </Overlay>
-      <Overlay
+      </ModalOverlay>
+      <ModalOverlay
         title={t("information.counterParty.title")}
         onClose={() => toggleShowOrderTypeInfo(false)}
         isHidden={!showOrderTypeInfo}
       >
         <OrderTypesModal onCloseButtonClick={() => toggleShowOrderTypeInfo()} />
-      </Overlay>
+      </ModalOverlay>
       <ProtocolFeeOverlay
         isHidden={showFeeInfo}
         onCloseButtonClick={() => toggleShowFeeInfo(false)}
