+++ src/components/@widgets/WalletConnector/WalletConnector.tsx
@@ -18,7 +18,7 @@ import {
 import { buildGnosisSafeConnector } from "../../../web3-connectors/gnosis";
 import { tryActivateConnector } from "../../../web3-connectors/helpers";
 import { WalletProvider } from "../../../web3-connectors/walletProviders";
-import Overlay from "../../Overlay/Overlay";
+import ModalOverlay from "../../ModalOverlay/ModalOverlay";
 import { StyledWalletProviderList } from "../SwapWidget/SwapWidget.styles";
 
 interface WalletConnectorProps {
@@ -126,7 +126,7 @@ const WalletConnector: FC<WalletConnectorProps> = ({ className }) => {
   );
 
   return (
-    <Overlay
+    <ModalOverlay
       isHidden={!showWalletList}
       title={t("wallet.selectWallet")}
       onClose={handleCloseButtonClick}
@@ -136,7 +136,7 @@ const WalletConnector: FC<WalletConnectorProps> = ({ className }) => {
         onClose={handleCloseButtonClick}
         onWalletProviderButtonClick={handleWalletProviderButtonClick}
       />
-    </Overlay>
+    </ModalOverlay>
   );
 };
