+++ src/components/InformationModals/subcomponents/OrderTypesModal/OrderTypesModal.styles.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components/macro";
 
-import { OverlayActionButton } from "../../../Overlay/Overlay.styles";
+import { OverlayActionButton } from "../../../ModalOverlay/ModalOverlay.styles";
 
 export const Container = styled.div`
   padding-bottom: 2rem;
